Initial implementation of exception handling proposal
There's not a ton of spec tests or reference tests in wabt for this yet,
so it still feels quite experimental. That being said this should parse
and be compliant (ish) with the [current state of the proposal][enxp]
except for the name section for exceptions, but that's just because it
doesn't seem too too important yet.

Closes #17

[exnp]: https://github.com/webassembly/exception-handling