Implement (ref T) and (struct (field valtype)..)
This commit extends value type to contain (ref T), and type to define
structs.

There are two large and slightly entangled changes.
 (1) ValType::Ref(Index) which requires ValType's to have their names
     resolved.
 (2) Type contains TypeDef which may be FunctionType or StructType.

The registration pass of resolve::names is extended to register field
names which are globally unique, but resolve to a relative index to the
struct they are defined on.

Every Ast item that contains a ValType is updated to resolve the
potential (ref T) they contain. This includes Types, Globals, Funcs,
Select, Block, If, Loop, CallIndirect.

TypeUse is in a bit of an ambiguous place as it's only used for
referencing/declaring a function type. There is no way to define a
struct type inline. The name might need to be updated down the road, but
that should probably wait for an actual GC spec to be developed.

The remaining GC proposal instructions are also implemented and are
fairly straightforward.