Make ref.null/is_null use a ref-type immediate
The type uses the formal reference-types production, so it's `ref.null
any` not `ref.null anyref`.

The elem segment parser also had to be updated to check the type of
ref.null used is compatible with the element segment type. This is
tricky with subtypes, so for now it just does strict equality.